Use ns.bench.v2 + Comment out every files that do not use it
