+++ bench/function/simd/plus.cpp
@@ -6,36 +6,17 @@
 //                            http://www.boost.org/LICENSE_1_0.txt
 // -------------------------------------------------------------------------------------------------
 
-#include <ns.bench.hpp>
+#include <simd_bench.hpp>
 #include <boost/simd/function/simd/plus.hpp>
 #include <boost/simd/pack.hpp>
 #include <cmath>
 
-namespace bs = boost::simd;
 namespace nsb = ns::bench;
 
-template <typename T>
-struct plus_simd
+DEFINE_SIMD_BENCH(simd_plus, boost::simd::plus);
+
+DEFINE_BENCH_MAIN()
 {
-   template <typename U>
-   void operator()(U min0, U max0, U min1, U max1)
-   {
-     using pack_t = bs::pack<T>;
-     using ret_type = bs::pack<T>;
-     nsb::make_function_experiment_cpe_sized_<pack_t::static_size>
-       ( [](const pack_t & x0, const pack_t & x1) -> ret_type
-         { return bs::plus(x0, x1); }
-       , nsb::generators::rand<pack_t>(min0, max0)
-       , nsb::generators::rand<pack_t>(min1, max1)
-       );
-   }
-};
-
-
-int main(int argc, char **argv) {
-   nsb::parse_args(argc, argv);
-   nsb::make_for_each<plus_simd, NS_BENCH_SIGNED_NUMERIC_TYPES>( -10,  10,  -10,  10);
-   nsb::make_for_each<plus_simd, NS_BENCH_UNSIGNED_NUMERIC_TYPES>(0,  10, 0,  10);
-   return 0;
+  nsb::for_each<simd_plus, NS_BENCH_SIGNED_NUMERIC_TYPES>(-10, 10, -10, 10);
+  nsb::for_each<simd_plus, NS_BENCH_UNSIGNED_NUMERIC_TYPES>(0, 10,  0,  10);
 }
-
