+++ bench/simd_bench.hpp
@@ -10,7 +10,7 @@
 #ifndef SIMD_BENCH_HPP_INCLUDED
 #define SIMD_BENCH_HPP_INCLUDED
 
-#include <ns.bench.v2.hpp>
+#include <ns.bench.hpp>
 #include <boost/config.hpp>
 #include <boost/simd/pack.hpp>
 #include <boost/preprocessor/cat.hpp>
